CAS - fixups for version 6
Note: I deleted the log4j1 configuration as log4j2 is in use now.
I also updated the template for the SP so that the ticket validation
once authenticated onto CAS is using the http connector directly onto
localhost. The other option would have been to trust the self-signed
certificate generated for the apache2 used as front webserver.