removing tomcat ssl connector on proxycas instance (#115)
Note: there are still some occurences for "8443" in the repository, but
related to the redirectPort, and even if I am not sure if this is still
needed, I think it comes from the default tomcat configuration. I am
pretty sure we can leave them as they are, as it won't harm. In the
other tomcat configurations, they appear in a commented block.

Also I kept all the logic of trusting the generated self signed
certificate, because I think it should be still in use by the JVM /
server-to-server communication.